+++ pybop/optimisers/_adamw.py
@@ -2,6 +2,8 @@
 # Extends the Pints' Adam Class with a Weight Decay addition
 #
 
+import warnings
+
 import numpy as np
 from pints import Optimiser as PintsOptimiser
 
@@ -55,7 +57,11 @@ class AdamWImpl(PintsOptimiser):
 
     def __init__(self, x0, sigma0=0.015, boundaries=None):
         if boundaries is not None:
-            print("NOTE: Boundaries ignored by AdamW")
+            warnings.warn(
+                "Boundaries ignored by AdamW",
+                UserWarning,
+                stacklevel=2,
+            )
 
         self.boundaries = None
         super().__init__(x0, sigma0, self.boundaries)